pyproject.toml
[tool.poetry]
name = "chores"
version = "0.1.0"
description = ""
authors = ["cappenz <charlotte@appenzeller.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
vapi-python = "^0.1.9"
openai = "^1.45.0"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
